### 2020-01-22 #2036 @whedon
[PRE REVIEW]: osfr: An R Interface to the Open Science Framework
**Submitting author:** @aaronwolen (<a href="http://orcid.org/0000-0003-2542-2202">Aaron R Wolen</a>)
**Repository:** <a href="https://github.com/ropensci/osfr" target ="_blank">https://github.com/ropensci/osfr</a>
**Version:** v0.2.7
