### 2020-01-22 #2036 @labarba
👋 @trallard — Could you handle this JOSS submission as editor? Let us know! Thx.
