### 2020-01-23 #2036 @aaronwolen
Thanks, @whedon. You rock.

This package has been reviewed by rOpenSci: https://github.com/ropensci/software-review/issues/279, @ropensci/editors
