### 2020-02-04 #2036 @maelle
This package has been reviewed by rOpenSci over at https://github.com/ropensci/software-review/issues/279 so doesn't need any review? @trallard
