### 2020-02-04 #2036 @kthyng
> This package has been reviewed by rOpenSci over at [ropensci/software-review#279](https://github.com/ropensci/software-review/issues/279) so doesn't need any review? @trallard

Hi @maelle! Ah, we must have missed that sorry. Just to verify: this submission has already been reviewed and accepted by rOpenSci?
