### 2020-02-04 #2036 @maelle
Yes, it has been reviewed and accepted by rOpenSci indeed.
