### 2020-02-04 #2036 @kthyng
Another link to the [rOpenSci](https://github.com/ropensci/software-review/issues/279) review.
