### 2020-02-04 #2071 @kthyng
This has already been reviewed and accepted by rOpenSci. Issue [#279](https://github.com/ropensci/software-review/issues/279)
