### 2020-02-04 #2071 @kthyng
@aaronwolen and @maelle I can't find your paper on a quick glance so I'll just ask you — what is with the "AND" in a bunch of the references? Is it meant to be that way or is that incorrectly forced capitalization?
