### 2020-02-04 #2071 @aaronwolen
The latter. I'm not sure how/why it happened but I'll fix in the repo. Thanks for catching it, @kthyng!
