### 2020-02-04 #2071 @aaronwolen
I fixed the capitalization issue in ropensci/osfr@d03f5bc.
