### 2020-02-04 #2071 @kthyng
> Also, the paper is located [here](https://github.com/ropensci/osfr/blob/master/inst/paper.md).

The paper is being found correctly by whedon right? As in, the last proof shows up the way you want? It looks good to me.
