### 2020-02-05 #2036 @MilesMcBain
Hi @trallard, Sorry I'm not familiar with this osf.io framework and do not have the bandwidth to immerse myself in it for a review at the moment.

There are some active R/rOpensci community members I know of in the communities listed here: https://osf.io/be7yt/wiki/home/ I am sure you could find someone suitable without much digging, e.g.: https://morn.netlify.com/about/
