### 2020-02-05 #2071 @aaronwolen
Thanks, @kthyng. The Zenodo archive for osfr is here: https://zenodo.org/record/3625248#.XjrSDhdMF24.

I just made a small tweak to my affiliation in the paper but otherwise the proof looks good to me 👍.
