### 2020-02-05 #2036 @kthyng
@MilesMcBain and @benmarwick sorry for the confusion — this has already been reviewed through rOpenSci so no reviewers needed here. Thanks!
