### 2020-02-05 #2071 @aaronwolen
Thanks, I updated the info but Zenodo throws the following error when I attempt to publish:

```error
Internal Server Error Error ID: f7916d075dfa44e993b9f48d0203b19e
```

I'm looking into it.
