### 2020-02-06 #2071 @aaronwolen
One question: should the *Reviewers* list also include the ROpenSci reviewers?

cc/@maelle
