### 2020-02-06 #2071 @arfon
@aaronwolen on the paper you mean? Currently the JOSS papers simply include the JOSS reviewers, including the rOpenSci reviewers in the PDF isn't supported sorry.
